requirements.txt
Foundation==0.1.0a0.dev1
matplotlib==3.10.8
mediapipe_silicon==0.9.2.1
numpy==2.4.1
opencv_contrib_python==4.13.0.90
opencv_python==4.13.0.90
pyobjc_core==12.1
pyobjc_framework_cocoa==12.1
pyobjc_framework_iobluetooth==12.1
pyobjc_framework_quartz==12.1
Quartz==0.0.1.dev0
